Fix GCC -Wconversion warnings

+++ finance/src/sgpp/finance/application/BlackScholesHullWhiteSolver.cpp
@@ -77,7 +77,7 @@ BlackScholesHullWhiteSolver::~BlackScholesHullWhiteSolver() {
 
 void BlackScholesHullWhiteSolver::constructGrid(base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
-  this->levels = level;
+  this->levels = static_cast<size_t>(level);
 
   this->myGrid = new base::LinearBoundaryGrid(BoundingBox);
